+++ src/angular-maps/services/geotxt.parser.ts
@@ -1,0 +1,18 @@
+import {LatLng} from 'leaflet';
+import {GeoElement, GeoElementType, GeoParser, LatLngTime} from './geo.parser';
+import {AbstractGeoTxtParser} from '@dps/mycms-commons/dist/geo-commons/services/geotxt.parser';
+import {GeoUtils} from './geo.utils';
+
+export class GeoTxtParser extends AbstractGeoTxtParser<LatLng> implements GeoParser {
+    protected createLatLng(lat: string | number, lng: string | number, alt?: number, time?: Date): LatLng {
+        return GeoUtils.createLatLng(lat, lng, alt, time);
+    }
+
+    protected createGeoElement(type: GeoElementType, points: LatLng[], name: string): GeoElement {
+        return GeoUtils.createGeoElement(type, points, name);
+    }
+
+    protected calcDistance(from: LatLng, to: LatLng): number {
+        return from.distanceTo(to);
+    }
+}